Remove require cycle
